Se realiza sincronización de la parte que valida si es usuario admin

+++ app/Http/Controllers/AuditorioController.php
@@ -32,8 +32,15 @@ class AuditorioController extends Controller
     public function create()
     {
         //
+        // me valida si soy un usuario administrador
 
-        return view('/layouts.super_admin.crearAuditorio');
+        if(Auth::user()->isAdmin()==true){
+            return view('/layouts.super_admin.crearAuditorio');
+          		            
+        }else{
+            return view('mensajeDeError');
+        }
+        
     }
 
     /**
@@ -76,8 +83,13 @@ class AuditorioController extends Controller
     public function edit($id)
     {
         //
-        $auditorio = Auditorio::find($id);
-		return view('layouts.super_admin.editarAuditorio',['auditorio'=>$auditorio]);
+        if(Auth::user()->isAdmin()==true){
+            $auditorio = Auditorio::find($id);
+		return view('layouts.super_admin.editarAuditorio',['auditorio'=>$auditorio]);         
+          		            
+        }else{
+            return view('mensajeDeError');
+        }        
 
     }
 
@@ -91,12 +103,18 @@ class AuditorioController extends Controller
     public function update(Request $request, $id)
     {
         //
-        $auditorio = Auditorio::find($id);
-		$auditorio->fill($request->all());
-		$auditorio->save();
+        if(Auth::user()->isAdmin()==true){
 
-		Session::flash('message','Auditorio editado Correctamente');
-		return Redirect::to('auditorios');
+            $auditorio = Auditorio::find($id);
+            $auditorio->fill($request->all());
+            $auditorio->save();
+
+            Session::flash('message','Auditorio editado Correctamente');
+            return Redirect::to('auditorios');
+
+        }else{
+            return view('mensajeDeError');
+        }
     }
 
     /**
@@ -106,9 +124,15 @@ class AuditorioController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy($id)
-    {
-        Auditorio::destroy($id);
-		Session::flash('message','Auditorio eliminado de manera correcta');
-		return Redirect::to('auditorios');
+    {       
+        //
+        if(Auth::user()->isAdmin()==true){
+            Auditorio::destroy($id);
+            Session::flash('message','Auditorio eliminado de manera correcta');
+            return Redirect::to('auditorios');
+          		            
+        }else{
+            return view('mensajeDeError');
+        }
     }
 }
